PICKED FROM TRUNK 21060
fix for SCHEDULING-1421

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/scheduling/branches/Scheduling_3.1.x@21062 28e8926c-6b08-0410-baaa-805c5e19b8d6

+++ extensions/scilab/PAscheduler/macros/PAResult_clean.sci
@@ -1,22 +1,25 @@
 function PAResult_clean(R)
+    if ~jexists(R.cleaned) then
+        return;
+    end
     if ~jinvoke(R.cleaned,'get')
-    warning('off')
-    setf = R.cleanFileSet;
-    for i=1:length(setf)
-        mdelete(setf(i));
-    end
-    setd = R.cleanDirSet;
-    for i=1:length(setd)
-        d=ls(setd(i));
-        if (getos() == "Linux") & size(d,1) == 2
-            rmdir(setd(i));
-        elseif (getos() == "Windows") & size(d,1) == 0
-            rmdir(setd(i));
+        warning('off')
+        setf = R.cleanFileSet;
+        for i=1:length(setf)
+            mdelete(setf(i));
         end
+        setd = R.cleanDirSet;
+        for i=1:length(setd)
+            d=ls(setd(i));
+            if (getos() == "Linux") & size(d,1) == 2
+                rmdir(setd(i));
+            elseif (getos() == "Windows") & size(d,1) == 0
+                rmdir(setd(i));
+            end
+        end
+        //sched = PAScheduler;
+        //sched.PATaskRepository(R.jobid, R.taskid, 'received');
+        jinvoke(R.cleaned,'set',%t);
+        warning('on')
     end
-    //sched = PAScheduler;
-    //sched.PATaskRepository(R.jobid, R.taskid, 'received');
-    jinvoke(R.cleaned,'set',%t);
-    warning('on')
-end
 endfunction
